feat: modify logic foreign_key by Django
