+++ src/bk-user/bkuser/apps/tenant/constants.py
@@ -8,8 +8,11 @@ Unless required by applicable law or agreed to in writing, software distributed 
 an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
+import pytz
 from blue_krill.data_types.enum import FeatureFlag, FeatureFlagField
 from django.utils.translation import gettext_lazy as _
+
+TIME_ZONE_CHOICES = [(i, i) for i in list(pytz.common_timezones)]
 
 
 class TenantFeatureFlag(FeatureFlag):  # type: ignore
